requirements.txt
flask==3.0.3
flask-cors==4.0.1
supabase==2.5.3
python-dotenv==1.0.1
python-dateutil==2.9.0
reportlab==4.2.5
Pillow==10.4.0
gunicorn==22.0.0
requests==2.32.3
PyJWT==2.8.0
bcrypt==4.2.0
edge-tts==6.1.12
